+++ src/android/com/outsystems/plugins/healthfitness/CordovaImplementation.kt
@@ -65,4 +65,6 @@ abstract  class CordovaImplementation : CordovaPlugin(), AndroidPlatformInterfac
         }
     }
 
+    abstract override fun areGooglePlayServicesAvailable(callbackContext: CallbackContext): Boolean
+
 }